Consistently pass float
PHP would do this for us, but this is a limitation of TransPHPile

+++ src/Phan/Phan.php
@@ -110,7 +110,7 @@ class Phan implements IgnoredFilesFilterInterface {
         $code_base->setCurrentParsedFile(null);
         foreach ($file_path_list as $i => $file_path) {
             $code_base->setCurrentParsedFile($file_path);
-            CLI::progress('parse', ($i + 1) / $file_count);
+            CLI::progress('parse', (float)(($i + 1) / $file_count));
 
             // Kick out anything we read from the former version
             // of this file
@@ -243,7 +243,7 @@ class Phan implements IgnoredFilesFilterInterface {
         // This worker takes a file and analyzes it
         $analysis_worker = function($i, $file_path)
             use ($file_count, $code_base, $temporary_file_mapping) {
-                CLI::progress('analyze', ($i + 1) / $file_count);
+                CLI::progress('analyze', (float)(($i + 1) / $file_count));
                 Analysis::analyzeFile($code_base, $file_path, $temporary_file_mapping[$file_path] ?? null);
             };
 
@@ -340,7 +340,7 @@ class Phan implements IgnoredFilesFilterInterface {
 
         CLI::progress('dependencies', 0.0);  // trigger UI update of 0%
         foreach ($file_path_list as $i => $file_path) {
-            CLI::progress('dependencies', ($i + 1) / $file_count);
+            CLI::progress('dependencies', (float)(($i + 1) / $file_count));
 
             // Add the file itself to the list
             $dependency_file_path_list[] = $file_path;
